+++ src/common/Pager/index.js
@@ -1,29 +1,16 @@
-import { useEffect, useState } from "react";
 import { useReplaceURLParameters } from "../../useURLParameters";
 import {
   Arrow,
-  ArrowDisabled,
   Button,
   ButtonText,
   ButtonWrapper,
   LetfArrow,
-  LetfArrowDisabled,
   Pages,
   PagesWrapper,
 } from "./styled";
 
 export const Pager = ({ page, totalPages }) => {
-  const [width, setWidth] = useState(window.innerWidth);
-  const breakpoint = 767;
   const replaceURLParameters = useReplaceURLParameters();
-
-  useEffect(() => {
-    const handleResizeWindow = () => setWidth(window.innerWidth);
-    window.addEventListener("resize", handleResizeWindow);
-    return () => {
-      window.removeEventListener("resize", handleResizeWindow);
-    };
-  }, []);
 
   const changePage = (page) => {
     replaceURLParameters([
@@ -38,18 +25,13 @@ export const Pager = ({ page, totalPages }) => {
     <ButtonWrapper>
       <>
         <Button disabled={page === 1} onClick={() => changePage(1)}>
-          {page === 1 ? <LetfArrowDisabled /> : <LetfArrow />}
-          {width > breakpoint ? (
-            <ButtonText>First</ButtonText>
-          ) : page === 1 ? (
-            <LetfArrowDisabled />
-          ) : (
-            <LetfArrow />
-          )}
+          <LetfArrow disabled={page === 1} />
+          <LetfArrow disabled={page === 1} />
+          <ButtonText>First</ButtonText>
         </Button>
         <Button disabled={page === 1} onClick={() => changePage(page - 1)}>
-          {page === 1 ? <LetfArrowDisabled /> : <LetfArrow />}
-          {width > breakpoint ? <ButtonText>Previous</ButtonText> : ""}
+          <LetfArrow disabled={page === 1} />
+          <ButtonText disabled={page === 1}>Previous</ButtonText>
         </Button>
       </>
 
@@ -64,21 +46,16 @@ export const Pager = ({ page, totalPages }) => {
           disabled={page === totalPages}
           onClick={() => changePage(page + 1)}
         >
-          {width > breakpoint ? <ButtonText>Next</ButtonText> : ""}
-          {page === totalPages ? <ArrowDisabled /> : <Arrow />}
+          <ButtonText disabled={page === totalPages}>Next</ButtonText>
+          <Arrow disabled={page === totalPages} />
         </Button>
         <Button
           disabled={page === totalPages}
           onClick={() => changePage(totalPages)}
         >
-          {width > breakpoint ? (
-            <ButtonText>Last</ButtonText>
-          ) : page === totalPages ? (
-            <ArrowDisabled />
-          ) : (
-            <Arrow />
-          )}
-          {page === totalPages ? <ArrowDisabled /> : <Arrow />}
+          <ButtonText disabled={page === totalPages}>Last</ButtonText>
+          <Arrow disabled={page === totalPages} />
+          <Arrow disabled={page === totalPages} />
         </Button>
       </>
     </ButtonWrapper>
